requirements.txt
numpy
Pillow
tqdm
torch
torch-hd
matplotlib
pyhdc
